inference/requirements.txt
torch>=2.0.0
transformers>=4.36.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
numpy>=1.24.0
psutil>=5.9.0
